+++ src/main/webapp/app/shared/export/csv-export-row-builder.ts
@@ -1,4 +1,4 @@
-import { ExportRowBuilder } from 'app/course/course-scores/export-row-builder';
+import { ExportRowBuilder } from 'app/shared/export/export-row-builder';
 import { CsvDecimalSeparator } from 'app/shared/export/export-modal.component';
 import { round } from 'app/shared/util/utils';
