+++ src/test/javascript/spec/component/course/course-scores/course-scores-excel-row-builder.spec.ts
@@ -1,7 +1,7 @@
 import { POINTS_KEY, SCORE_KEY } from 'app/shared/export/export-constants';
 import { ExerciseType } from 'app/entities/exercise.model';
-import { ExportRowBuilder } from 'app/course/course-scores/export-row-builder';
-import { ExcelExportRowBuilder } from 'app/course/course-scores/excel-export-row-builder';
+import { ExportRowBuilder } from 'app/shared/export/export-row-builder';
+import { ExcelExportRowBuilder } from 'app/shared/export/excel-export-row-builder';
 
 describe('The ExcelExportRowBuilder', () => {
     let excelRowBuilder: ExportRowBuilder;
